refactor(Lotto.js, LottoController.js): Lotto 클래스에 ValidateWinningNumbers를 주입하고 검증 로직 재사용
- Lotto 클래스에 validator 필드 생성
- ValidateWinningNumbers를 활용하여 Lotto 클래스에서 검증
- Lotto.js를 /src로 이동
- LottoController에서 진행되었던 ValidateWinningNumbers 검증 제거
- LottoController에서 에러 발생 시 error.message 출력
