+++ src/Lotto.js
@@ -1,15 +1,20 @@
+import ValidateWinningNumbers from './models/ValidateWinningNumbers.js';
+
 class Lotto {
   #numbers;
 
+  #validator;
+
   constructor(numbers) {
+    this.#validator = new ValidateWinningNumbers();
     this.#validate(numbers);
     this.#numbers = numbers.sort((a, b) => a - b);
   }
 
   #validate(numbers) {
-    if (numbers.length !== 6) {
-      throw new Error('[ERROR] 로또 번호는 6개여야 합니다.');
-    }
+    this.#validator.validateWinningNumbersFormat(numbers.join(','));
+    this.#validator.validateDuplicateNumbers(numbers);
+    numbers.forEach((number) => this.#validator.validateNumberRange(number));
   }
 
   getNumbers() {
